requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
skia-python>=144.0
pillow>=11.0
python-multipart>=0.0.9
numpy>=2.0
pytest>=8.0
httpx>=0.27
